requirements.txt
cache_to_disk
python-dotenv
python-slugify
python-twitter
feedparser
feedsearch